# backend/tests/contract/openapi.json
{
 "components": {
  "schemas": {
...
   "Benchmark": {
    "properties": {
     "currency": {
      "anyOf": [
       {
        "$ref": "#/components/schemas/Currency"
       },
       {
        "type": "null"
       }
      ]
     },
     "currency_id": {
      "anyOf": [
       {
        "type": "integer"
       },
       {
        "type": "null"
       }
      ],
      "title": "Currency Id"
     },
     "id": {
      "title": "Id",
      "type": "integer"
     },
     "name": {
      "title": "Name",
      "type": "string"
     },
     "short_name": {
      "title": "Short Name",
      "type": "string"
     },
     "symbol": {
      "title": "Symbol",
      "type": "string"
     }
    },
    "required": [
     "id",
     "name",
     "short_name",
     "symbol"
    ],
    "title": "Benchmark",
    "type": "object"
   },
...
   "CreatePortfolioRequest": {
    "properties": {
     "name": {
      "title": "Name",
      "type": "string"
     },
     "user_categories": {
      "items": {
       "$ref": "#/components/schemas/app__modules__portfolio__api__portfolio__schemas__CustomCategory-Input"
      },
      "title": "User Categories",
      "type": "array"
     }
    },
    "required": [
     "name",
     "user_categories"
    ],
    "title": "CreatePortfolioRequest",
    "type": "object"
   },
   "Currency": {
    "properties": {
     "code": {
      "title": "Code",
      "type": "string"
     },
     "id": {
      "title": "Id",
      "type": "integer"
     },
     "name": {
      "title": "Name",
      "type": "string"
     }
    },
    "required": [
     "id",
     "code",
     "name"
    ],
    "title": "Currency",
    "type": "object"
   },
   "CustomCategory-Output": {
    "properties": {
     "benchmark": {
      "anyOf": [
       {
        "$ref": "#/components/schemas/Benchmark"
       },
       {
        "type": "null"
       }
      ]
     },
     "benchmark_id": {
      "title": "Benchmark Id",
      "type": "integer"
     },
     "color": {
      "title": "Color",
      "type": "string"
     },
     "id": {
      "anyOf": [
       {
        "type": "integer"
       },
       {
        "type": "null"
       }
      ],
      "title": "Id"
     },
     "name": {
      "title": "Name",
      "type": "string"
     },
     "portfolio_id": {
      "anyOf": [
       {
        "type": "integer"
       },
       {
        "type": "null"
       }
      ],
      "title": "Portfolio Id"
     }
    },
    "required": [
     "name",
     "color",
     "benchmark_id"
    ],
    "title": "CustomCategory",
    "type": "object"
   },
...
   "HTTPValidationError": {
    "properties": {
     "detail": {
      "items": {
       "$ref": "#/components/schemas/ValidationError"
      },
      "title": "Detail",
      "type": "array"
     }
    },
    "title": "HTTPValidationError",
    "type": "object"
   },
...
   "Portfolio": {
    "properties": {
     "custom_categories": {
      "default": [],
      "items": {
       "$ref": "#/components/schemas/CustomCategory-Output"
      },
      "title": "Custom Categories",
      "type": "array"
     },
     "id": {
      "title": "Id",
      "type": "integer"
     },
     "name": {
      "title": "Name",
      "type": "string"
     },
     "user_id": {
      "title": "User Id",
      "type": "integer"
     }
    },
    "required": [
     "id",
     "name",
     "user_id"
    ],
    "title": "Portfolio",
    "type": "object"
   },
   "PortfolioSummary": {
    "description": "Portfolio identity without its categories, for administrative listings.",
    "properties": {
     "id": {
      "title": "Id",
      "type": "integer"
     },
     "name": {
      "title": "Name",
      "type": "string"
     },
     "user_id": {
      "title": "User Id",
      "type": "integer"
     }
    },
    "required": [
     "id",
     "name",
     "user_id"
    ],
    "title": "PortfolioSummary",
    "type": "object"
   },
...
   "UpdatePortfolioRequest": {
    "properties": {
     "id": {
      "title": "Id",
      "type": "integer"
     },
     "name": {
      "title": "Name",
      "type": "string"
     },
     "user_categories": {
      "anyOf": [
       {
        "items": {
         "$ref": "#/components/schemas/app__modules__portfolio__api__portfolio__schemas__CustomCategory-Input"
        },
        "type": "array"
       },
       {
        "type": "null"
       }
      ],
      "title": "User Categories"
     }
    },
    "required": [
     "id",
     "name"
    ],
    "title": "UpdatePortfolioRequest",
    "type": "object"
   },
...
   "ValidationError": {
    "properties": {
     "loc": {
      "items": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "integer"
        }
       ]
      },
      "title": "Location",
      "type": "array"
     },
     "msg": {
      "title": "Message",
      "type": "string"
     },
     "type": {
      "title": "Error Type",
      "type": "string"
     }
    },
    "required": [
     "loc",
     "msg",
     "type"
    ],
    "title": "ValidationError",
    "type": "object"
   },
...
   "app__modules__portfolio__api__portfolio__schemas__CustomCategory-Input": {
    "properties": {
     "benchmark_id": {
      "title": "Benchmark Id",
      "type": "integer"
     },
     "color": {
      "title": "Color",
      "type": "string"
     },
     "id": {
      "anyOf": [
       {
        "type": "integer"
       },
       {
        "type": "null"
       }
      ],
      "title": "Id"
     },
     "name": {
      "title": "Name",
      "type": "string"
     },
     "portfolio_id": {
      "anyOf": [
       {
        "type": "integer"
       },
       {
        "type": "null"
       }
      ],
      "title": "Portfolio Id"
     }
    },
    "required": [
     "name",
     "color",
     "benchmark_id"
    ],
    "title": "CustomCategory",
    "type": "object"
   }
  },
  "securitySchemes": {
   "OAuth2PasswordBearer": {
    "flows": {
     "password": {
      "scopes": {},
      "tokenUrl": "auth/jwt/login"
     }
    },
    "type": "oauth2"
   }
  }
 },
 "info": {
  "description": "API for portfolio consolidation",
  "title": "My Stonks API",
  "version": "1.0.0"
 },
 "openapi": "3.1.0",
 "paths": {
...
  "/portfolio": {
   "get": {
    "operationId": "list_user_portfolios_portfolio_get",
    "responses": {
     "200": {
      "content": {
       "application/json": {
        "schema": {
         "items": {
          "$ref": "#/components/schemas/Portfolio"
         },
         "title": "Response List User Portfolios Portfolio Get",
         "type": "array"
        }
       }
      },
      "description": "Successful Response"
     }
    },
    "security": [
     {
      "OAuth2PasswordBearer": []
     }
    ],
    "summary": "List User Portfolios"
   },
   "post": {
    "operationId": "create_portfolio_portfolio_post",
    "requestBody": {
     "content": {
      "application/json": {
       "schema": {
        "$ref": "#/components/schemas/CreatePortfolioRequest"
       }
      }
     },
     "required": true
    },
    "responses": {
     "200": {
      "content": {
       "application/json": {
        "schema": {}
       }
      },
      "description": "Successful Response"
     },
     "422": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      },
      "description": "Validation Error"
     }
    },
    "security": [
     {
      "OAuth2PasswordBearer": []
     }
    ],
    "summary": "Create Portfolio"
   }
  },
  "/portfolio/all": {
   "get": {
    "description": "Every portfolio in the application, for administrative screens.",
    "operationId": "list_all_portfolios_portfolio_all_get",
    "responses": {
     "200": {
      "content": {
       "application/json": {
        "schema": {
         "items": {
          "$ref": "#/components/schemas/PortfolioSummary"
         },
         "title": "Response List All Portfolios Portfolio All Get",
         "type": "array"
        }
       }
      },
      "description": "Successful Response"
     }
    },
    "security": [
     {
      "OAuth2PasswordBearer": []
     }
    ],
    "summary": "List All Portfolios"
   }
  },
...
  "/portfolio/position_consolidator/{portfolio_id}/consolidate": {
   "post": {
    "operationId": "consolidate_portfolio_portfolio_position_consolidator__portfolio_id__consolidate_post",
    "parameters": [
     {
      "in": "path",
      "name": "portfolio_id",
      "required": true,
      "schema": {
       "title": "Portfolio Id",
       "type": "integer"
      }
     }
    ],
    "responses": {
     "200": {
      "content": {
       "application/json": {
        "schema": {}
       }
      },
      "description": "Successful Response"
     },
     "422": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      },
      "description": "Validation Error"
     }
    },
    "security": [
     {
      "OAuth2PasswordBearer": []
     }
    ],
    "summary": "Consolidate Portfolio",
    "tags": [
     "Portfolio Position Consolidator"
    ]
   }
  },
  "/portfolio/position_consolidator/{portfolio_id}/consolidate_asset_type_returns": {
   "post": {
    "operationId": "consolidate_asset_type_returns_portfolio_position_consolidator__portfolio_id__consolidate_asset_type_returns_post",
    "parameters": [
     {
      "in": "path",
      "name": "portfolio_id",
      "required": true,
      "schema": {
       "title": "Portfolio Id",
       "type": "integer"
      }
     }
    ],
    "responses": {
     "200": {
      "content": {
       "application/json": {
        "schema": {}
       }
      },
      "description": "Successful Response"
     },
     "422": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      },
      "description": "Validation Error"
     }
    },
    "security": [
     {
      "OAuth2PasswordBearer": []
     }
    ],
    "summary": "Consolidate Asset Type Returns",
    "tags": [
     "Portfolio Position Consolidator"
    ]
   }
  },
  "/portfolio/position_consolidator/{portfolio_id}/consolidate_category_returns": {
   "post": {
    "operationId": "consolidate_category_returns_portfolio_position_consolidator__portfolio_id__consolidate_category_returns_post",
    "parameters": [
     {
      "in": "path",
      "name": "portfolio_id",
      "required": true,
      "schema": {
       "title": "Portfolio Id",
       "type": "integer"
      }
     }
    ],
    "responses": {
     "200": {
      "content": {
       "application/json": {
        "schema": {}
       }
      },
      "description": "Successful Response"
     },
     "422": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      },
      "description": "Validation Error"
     }
    },
    "security": [
     {
      "OAuth2PasswordBearer": []
     }
    ],
    "summary": "Consolidate Category Returns",
    "tags": [
     "Portfolio Position Consolidator"
    ]
   }
  },
  "/portfolio/position_consolidator/{portfolio_id}/consolidate_portfolio_returns": {
   "post": {
    "operationId": "consolidate_portfolio_returns_portfolio_position_consolidator__portfolio_id__consolidate_portfolio_returns_post",
    "parameters": [
     {
      "in": "path",
      "name": "portfolio_id",
      "required": true,
      "schema": {
       "title": "Portfolio Id",
       "type": "integer"
      }
     }
    ],
    "responses": {
     "200": {
      "content": {
       "application/json": {
        "schema": {}
       }
      },
      "description": "Successful Response"
     },
     "422": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      },
      "description": "Validation Error"
     }
    },
    "security": [
     {
      "OAuth2PasswordBearer": []
     }
    ],
    "summary": "Consolidate Portfolio Returns",
    "tags": [
     "Portfolio Position Consolidator"
    ]
   }
  },
  "/portfolio/position_consolidator/{portfolio_id}/recalculate_all_position": {
   "post": {
    "operationId": "recalculate_all_positions_portfolio_position_consolidator__portfolio_id__recalculate_all_position_post",
    "parameters": [
     {
      "in": "path",
      "name": "portfolio_id",
      "required": true,
      "schema": {
       "title": "Portfolio Id",
       "type": "integer"
      }
     }
    ],
    "responses": {
     "200": {
      "content": {
       "application/json": {
        "schema": {}
       }
      },
      "description": "Successful Response"
     },
     "422": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      },
      "description": "Validation Error"
     }
    },
    "security": [
     {
      "OAuth2PasswordBearer": []
     }
    ],
    "summary": "Recalculate All Positions",
    "tags": [
     "Portfolio Position Consolidator"
    ]
   }
  },
  "/portfolio/position_consolidator/{portfolio_id}/recalculate_asset_position": {
   "post": {
    "operationId": "consolidate_portfolio_asset_portfolio_position_consolidator__portfolio_id__recalculate_asset_position_post",
    "parameters": [
     {
      "in": "path",
      "name": "portfolio_id",
      "required": true,
      "schema": {
       "title": "Portfolio Id",
       "type": "integer"
      }
     },
     {
      "in": "query",
      "name": "asset_id",
      "required": true,
      "schema": {
       "title": "Asset Id",
       "type": "integer"
      }
     }
    ],
    "responses": {
     "200": {
      "content": {
       "application/json": {
        "schema": {}
       }
      },
      "description": "Successful Response"
     },
     "422": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      },
      "description": "Validation Error"
     }
    },
    "security": [
     {
      "OAuth2PasswordBearer": []
     }
    ],
    "summary": "Consolidate Portfolio Asset",
    "tags": [
     "Portfolio Position Consolidator"
    ]
   }
  },
...
  "/portfolio/{portfolio_id}": {
   "delete": {
    "operationId": "delete_portfolio_portfolio__portfolio_id__delete",
    "parameters": [
     {
      "in": "path",
      "name": "portfolio_id",
      "required": true,
      "schema": {
       "title": "Portfolio Id",
       "type": "integer"
      }
     }
    ],
    "responses": {
     "200": {
      "content": {
       "application/json": {
        "schema": {}
       }
      },
      "description": "Successful Response"
     },
     "422": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      },
      "description": "Validation Error"
     }
    },
    "security": [
     {
      "OAuth2PasswordBearer": []
     }
    ],
    "summary": "Delete Portfolio"
   },
   "put": {
    "operationId": "update_portfolio_portfolio__portfolio_id__put",
    "parameters": [
     {
      "in": "path",
      "name": "portfolio_id",
      "required": true,
      "schema": {
       "title": "Portfolio Id",
       "type": "integer"
      }
     }
    ],
    "requestBody": {
     "content": {
      "application/json": {
       "schema": {
        "$ref": "#/components/schemas/UpdatePortfolioRequest"
       }
      }
     },
     "required": true
    },
    "responses": {
     "200": {
      "content": {
       "application/json": {
        "schema": {}
       }
      },
      "description": "Successful Response"
     },
     "422": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      },
      "description": "Validation Error"
     }
    },
    "security": [
     {
      "OAuth2PasswordBearer": []
     }
    ],
    "summary": "Update Portfolio"
   }
  },
...
 }
}